backend/requirements.txt
fastapi
uvicorn
pyautogui
websockets
pywebview
qrcode
pillow
jinja2
pystray
cryptography